refactor : manage page 리팩토링
