--- src/1_Page/Team/ui/HistoryListBox/type.d.ts
@@ -1,1 +1,0 @@
-
