+++ src/1_Page/Team/ui/ManagePage/type.d.ts
@@ -1,6 +1,4 @@
-import { TeamInfo } from "../../../../3_Entity/Team/types/response";
-
-export type ManagePageProps = {
+type ManagePageProps = {
   teamInfo: TeamInfo;
   handleTogglePage: () => void;
 };
