--- src/1_Page/Team/ui/ManagePage/ui/AutoMatchModalBtn/ui/lib/time.ts
@@ -1,5 +1,0 @@
-export const timesFor30 = Array.from({ length: 48 }, (_, i) => {
-  const hour = String(Math.floor(i / 2)).padStart(2, "0");
-  const minute = i % 2 === 0 ? "00" : "30";
-  return `${hour}:${minute}`;
-});
