+++ src/1_Page/Team/ui/ManagePage/ui/BannerImageInput/index.tsx
@@ -1,10 +1,7 @@
 import { useForm, SubmitHandler } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
-
-import { BannerImageInputProps, ImageForm } from "./type";
 import { schema } from "../../../../../../4_Shared/lib/imgSchema";
 import useImageHandler from "./model/useImageHandler";
-
 import usePutTeamBanner from "../../../../../../3_Entity/Team/usePutTeamBanner";
 
 const BannerImageInput = (props: BannerImageInputProps) => {
@@ -38,7 +35,6 @@ const BannerImageInput = (props: BannerImageInputProps) => {
 
   const onSubmit: SubmitHandler<ImageForm> = (props) => {
     handleSave();
-    if (!props.img) return;
     putTeamBanner(props.img);
   };
