+++ src/1_Page/Team/ui/ManagePage/ui/BannerImageInput/model/type.d.ts
@@ -1,7 +1,8 @@
 import { UseFormClearErrors, UseFormSetValue } from "react-hook-form";
+// import 구문으로 전역 사용 불가
 
 export type UseImageHandlerProps = {
   imgSrc: string;
-  setValue: UseFormSetValue<{ img: File | null }>;
+  setValue: UseFormSetValue<{ img: File }>;
   clearErrors: UseFormClearErrors<{ img: string }>;
 };
