+++ src/1_Page/Team/ui/ManagePage/ui/EmblemImageInput/model/type.d.ts
@@ -2,6 +2,6 @@ import { UseFormClearErrors, UseFormSetValue } from "react-hook-form";
 
 export type UseImageHandlerProps = {
   imgSrc: string;
-  setValue: UseFormSetValue<{ img: File | null }>;
+  setValue: UseFormSetValue<{ img: File }>;
   clearErrors: UseFormClearErrors<{ img: string }>;
 };
