+++ src/1_Page/Team/ui/ManagePage/ui/EmblemImageInput/type.d.ts
@@ -1,8 +1,8 @@
-export type ImageInputProps = {
+type ImageInputProps = {
   team_list_idx: number;
   imgSrc: string;
 };
 
-export type ImageForm = {
-  img: File | null;
+type ImageForm = {
+  img: File;
 };
