+++ src/1_Page/Team/ui/ManagePage/ui/TeamApplications/type.d.ts
@@ -1,3 +1,3 @@
-export type TeamApplicationsProps = {
+type TeamApplicationsProps = {
   team_list_idx: number;
 };
