+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/model/type.d.ts
@@ -1,5 +1,4 @@
 import { UseFormReset, UseFormSetValue } from "react-hook-form";
-import { TeamInfoForm } from "../type";
 
 export type UseManageModifyProps = {
   setValue: UseFormSetValue<TeamInfoForm>;
