+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/model/useManageModify.ts
@@ -1,6 +1,5 @@
 import React from "react";
 
-import { TeamInfoForm } from "../type";
 import { UseManageModifyProps } from "./type";
 
 const useManageModify = (
