+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/ui/InputFiled/index.tsx
@@ -1,5 +1,4 @@
 import { useFormContext } from "react-hook-form";
-import { InputFieldProps } from "./type";
 
 const InputField = (props: InputFieldProps) => {
   const { label, modifyMode, type = "text", name, placeholder = "" } = props;
