+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/ui/InputFiled/type.d.ts
@@ -1,4 +1,4 @@
-export type InputFieldProps = {
+type InputFieldProps = {
   label: string;
   name: string;
   modifyMode: boolean;
