+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/ui/StautsRadio/index.tsx
@@ -1,5 +1,4 @@
 import { useFormContext } from "react-hook-form";
-import { StatusRadioProps } from "./type";
 
 const StatusRadio = (props: StatusRadioProps) => {
   const { modifyMode } = props;
