+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/ui/StautsRadio/type.d.ts
@@ -1,1 +1,1 @@
-export type StatusRadioProps = { modifyMode: boolean };
+type StatusRadioProps = { modifyMode: boolean };
