+++ src/1_Page/Team/ui/ManagePage/ui/TextInputForm/ui/TeamNameInput/type.d.ts
@@ -1,4 +1,4 @@
-export type TeamNameCheckInputProps = {
+type TeamNameCheckInputProps = {
   modifyMode: boolean;
   isShort: boolean;
 };
