+++ src/4_Shared/lib/dateFormatter.ts
@@ -22,3 +22,9 @@ export const formatDateKoreanDate = (date: Date): string => {
     minute: "2-digit",
   });
 };
+
+export const formatTime30 = Array.from({ length: 48 }, (_, i) => {
+  const hour = String(Math.floor(i / 2)).padStart(2, "0");
+  const minute = i % 2 === 0 ? "00" : "30";
+  return `${hour}:${minute}`;
+});
